+++ activerecord/test/cases/adapters/postgresql/connection_test.rb
@@ -90,7 +90,7 @@ module ActiveRecord
     end
 
     def test_tables_logs_name
-      @connection.tables('hello')
+      ActiveSupport::Deprecation.silence { @connection.tables('hello') }
       assert_equal 'SCHEMA', @subscriber.logged[0][1]
     end
 
@@ -100,7 +100,7 @@ module ActiveRecord
     end
 
     def test_table_exists_logs_name
-      @connection.table_exists?('items')
+      ActiveSupport::Deprecation.silence { @connection.table_exists?('items') }
       assert_equal 'SCHEMA', @subscriber.logged[0][1]
     end